backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
PyMySQL==1.1.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.7.0
djangorestframework_simplejwt==5.5.0
cryptography==45.0.5
pillow==11.3.0
pymongo==4.13.2